Fix figure numbering is overriden when use the extension
** Why is this change required? **

When numfig is set to true, figure directives are automatically
numbered. However, before this fix, the caption of figures is
overriden and the automatically numbering is gone. This commit
fixed the issue (also listed on github
https://github.com/xxks-kkk/sphinxcontrib-pseudocode/issues/14).

** How does this change solve the issue? **

Instead of using nodes.caption directly (also with override when
add_node, which causes bug), we create a subclass pseudocodeCaption
and use it instead.

** Does this commit cause any expected behavior to change?  If so, what? **

** Issue: **

** Design Docs: **